events almost fully functional
